+++ binkapp/Screens/PaymentWallet/PaymentWalletViewModel.swift
@@ -43,6 +43,11 @@ class PaymentWalletViewModel: WalletViewModel {
 
     func showDeleteConfirmationAlert(card: CD_PaymentCard, yesCompletion: @escaping () -> Void, noCompletion: @escaping () -> Void) {
         router.showDeleteConfirmationAlert(withMessage: "delete_card_confirmation".localized, yesCompletion: { [weak self] in
+            guard Current.apiManager.networkIsReachable else {
+                self?.router.presentNoConnectivityPopup()
+                noCompletion()
+                return
+            }
             self?.repository.delete(card, completion: yesCompletion)
         }, noCompletion: {
             DispatchQueue.main.async {
